Organize metrics tests
We have placed several metrics-related tests in several files. It is a
bit hard to maintain and read it. So organize them.
